fix: move SynConfig to test package
#1387
#1196
